+++ src/MyPlugin.ts
@@ -43,44 +43,19 @@ export module MyPlugin {
 				const matrix = await buildMatrix(joplin,notes,columns,options.columns);
 				for(let col of matrix){ // Content
 					html += "<td style='vertical-align:top'>";
-					for(let note of col){
-						//const noteTags = await joplin.data.get(['notes',note.id,'tags']);
-						//console.log("NoteTags",noteTags);
-						html += await renderNote(joplin, note, uiOptions);
+					for(let note of col.items){
+						html += await renderNote(joplin, note, uiOptions,col.options);
 					}
 					html += "</td>";
 				}
 				html += "</tr>";
 				html += "</table>";
 			}
-			/*html += "<table>";
-			html += "<tr>";
-			for(let col of options.columns){ // Header
-				html += "<th>"+col.label;
-				if(col.tag){
-					html += "<br>("+col.tag+")";
-				}
-				html += "</th>";
-			}
-			html += "</tr>";
-			html += "<tr>";
-			const matrix = await buildMatrix(joplin,notes,options.columns);
-			for(let col of matrix){ // Content
-				html += "<td style='vertical-align:top'>";
-				for(let note of col){
-					//const noteTags = await joplin.data.get(['notes',note.id,'tags']);
-					//console.log("NoteTags",noteTags);
-					html += await renderNote(joplin, note, uiOptions);
-				}
-				html += "</td>";
-			}
-			html += "</tr>";
-			html += "</table>";*/
+			html += "<div>"+notes.items.length+" notes</div>";
 		}catch(err){
 			console.error(err);
 			html = "Error:"+err;
 		}
-
 		return html;
 	}
 }
@@ -93,6 +68,10 @@ returns [[{id:...},{},...],[...],...]
 async function buildMatrix(joplin,notes,columns,optColumns){
 	console.log("OC",optColumns);
 	let matrix = MyPluginMethods.initColumns(columns);
+	for(let i=0; i< matrix.length;i++){
+		matrix[i].items = [];
+		matrix[i].options = columns[i];
+	}
 	for(let note of notes.items){
 		console.log("Note",note);
 		const noteTags = await joplin.data.get(['notes',note.id,'tags']);
@@ -101,7 +80,7 @@ async function buildMatrix(joplin,notes,columns,optColumns){
 		console.log("Columns",colArray);
 
 		for(let colIndex of colArray){
-			matrix[colIndex].push(note);
+			matrix[colIndex].items.push(note);
 		}
 	}
 	console.log("Matrix:",matrix);
@@ -109,31 +88,53 @@ async function buildMatrix(joplin,notes,columns,optColumns){
 	//return [notes.items,[],[],[]];
 }
 
-async function renderNote(joplin, note, uiOptions){
-	const notedata = await joplin.data.get(['notes',note.id],{ fields: ['id', 'title','is_todo','todo_due','todo_completed'] });
+async function renderNote(joplin, note, uiOptions,colOption){
+	const notedata = await joplin.data.get(['notes',note.id],{ fields: ['id', 'title','is_todo','todo_due','todo_completed', 'updated_time'] });
 	/*
 	{	...
     "is_todo": 1/0,
     "todo_due": 0/1638385200000,
     "todo_completed": 0/1638385200000,
+		"updated_time": 0/1638385200000,
     ...
 	}
 	*/
 	console.log("Render note",notedata);
 
+	if(colOption.hideAfterXDaysDone && Number.isInteger(colOption.hideAfterXDaysDone)){
+		if(notedata.todo_completed != 0){ // 0 = not completed
+			const timeInfoCompleted = MyPluginMethods.getTimeDiff(notedata.todo_completed);
+			if(timeInfoCompleted.days >= colOption.hideAfterXDaysDone){
+				return ""; // => did not show the note
+			}
+		}
+	}
 
 	const onClick = uiOptions.noteOnClick(note).replace(/\n/g, ' ');
 
 	let uiCompleted = "";
 	if(notedata.todo_completed != 0){ // 0 = not completed
 		const timeInfoCompleted = MyPluginMethods.getTimeDiff(notedata.todo_completed);
-		uiCompleted = "<div style='font-style: italic;''>Done since "+timeInfoCompleted.ui+" </div>";
+		uiCompleted = "<div style='font-style: italic;'>Done since "+timeInfoCompleted.ui+" </div>";
 	}
 
+	let uiUpdated = "";
+	if(colOption.maxUnupdatedDaysTreshold && Number.isInteger(colOption.maxUnupdatedDaysTreshold)){
+		if(notedata.updated_time != 0){ // 0 = not updated
+			const timeInfoUpdated = MyPluginMethods.getTimeDiff(notedata.updated_time);
+			if(timeInfoUpdated.days >= colOption.maxUnupdatedDaysTreshold){
+				//uiUpdated = "<div style='color:red;font-style: italic;'>😴Last updated "+timeInfoUpdated.ui+" ago</div>";
+				uiUpdated = "<span title='Last updated "+timeInfoUpdated.ui+" ago'>😴</span>";
+			}
+		}
+	}
+
+
 	return `
-		<div style='padding:10px; margin-bottom: 10px; border: 1px solid;cursor:pointer;' title="${notedata.title}" onClick="${onClick}">
-			${notedata.title}
+		<div style='padding:10px; margin-bottom: 10px; border: 1px solid; border-radius: 7px; cursor:pointer;' title="${notedata.title}" onClick="${onClick}">
+			<div style='font-weight:bold;'>${notedata.title}</div>
 			${uiCompleted}
+			${uiUpdated}
 		</div>
 	`;
 }