+++ wiki/releasenotes.md
@@ -1,5 +1,9 @@
 # Releasenotes
 
+## Version 1.1.2
+* Parameter: maxUnupdatedDaysTreshold => Visual feedback if note is unedited longer than x days
+* Parameter: hideAfterXDaysDone => hide task if its done longer than X days ago
+
 ## Version 1.1.1
 * Label "Done since"